+++ platforms/cuda/src/kernels/kCalculateCDLJObcGbsaForces1.h
@@ -94,14 +94,14 @@ void METHOD_NAME(kCalculateCDLJObcGbsa, Forces1_kernel)(unsigned int* workUnit )
                     float dy                = psA[j].y - apos.y;
                     float dz                = psA[j].z - apos.z;
 #ifdef USE_PERIODIC
-                    dx                     -= floor(dx*cSim.invPeriodicBoxSizeX+0.5f)*cSim.periodicBoxSizeX;
-                    dy                     -= floor(dy*cSim.invPeriodicBoxSizeY+0.5f)*cSim.periodicBoxSizeY;
-                    dz                     -= floor(dz*cSim.invPeriodicBoxSizeZ+0.5f)*cSim.periodicBoxSizeZ;
+                    dx                     -= floorf(dx*cSim.invPeriodicBoxSizeX+0.5f)*cSim.periodicBoxSizeX;
+                    dy                     -= floorf(dy*cSim.invPeriodicBoxSizeY+0.5f)*cSim.periodicBoxSizeY;
+                    dz                     -= floorf(dz*cSim.invPeriodicBoxSizeZ+0.5f)*cSim.periodicBoxSizeZ;
 #endif
                     float r2                = dx * dx + dy * dy + dz * dz;
 
                     // CDLJ part
-                    float invR              = 1.0f / sqrt(r2);
+                    float invR              = 1.0f / sqrtf(r2);
                     float sig               = a.x + psA[j].sig;
                     float sig2              = invR * sig;
                     sig2                   *= sig2;
@@ -123,9 +123,9 @@ void METHOD_NAME(kCalculateCDLJObcGbsa, Forces1_kernel)(unsigned int* workUnit )
                     // ObcGbsaForce1 part
                     float alpha2_ij         = br * psA[j].br;
                     float D_ij              = r2 / (4.0f * alpha2_ij);
-                    float expTerm           = exp(-D_ij);
+                    float expTerm           = expf(-D_ij);
                     float denominator2      = r2 + alpha2_ij * expTerm;
-                    float denominator       = sqrt(denominator2);
+                    float denominator       = sqrtf(denominator2);
                     float Gpol              = (q2 * psA[j].q) / (denominator * denominator2);
                     float dGpol_dalpha2_ij  = -0.5f * Gpol * expTerm * (1.0f + D_ij);
                     dEdR                   += Gpol * (1.0f - 0.25f * expTerm);
@@ -164,14 +164,14 @@ void METHOD_NAME(kCalculateCDLJObcGbsa, Forces1_kernel)(unsigned int* workUnit )
                     float dy                = psA[j].y - apos.y;
                     float dz                = psA[j].z - apos.z;
 #ifdef USE_PERIODIC
-                    dx                     -= floor(dx*cSim.invPeriodicBoxSizeX+0.5f)*cSim.periodicBoxSizeX;
-                    dy                     -= floor(dy*cSim.invPeriodicBoxSizeY+0.5f)*cSim.periodicBoxSizeY;
-                    dz                     -= floor(dz*cSim.invPeriodicBoxSizeZ+0.5f)*cSim.periodicBoxSizeZ;
+                    dx                     -= floorf(dx*cSim.invPeriodicBoxSizeX+0.5f)*cSim.periodicBoxSizeX;
+                    dy                     -= floorf(dy*cSim.invPeriodicBoxSizeY+0.5f)*cSim.periodicBoxSizeY;
+                    dz                     -= floorf(dz*cSim.invPeriodicBoxSizeZ+0.5f)*cSim.periodicBoxSizeZ;
 #endif
                     float r2                = dx * dx + dy * dy + dz * dz;
 
                     // CDLJ part
-                    float invR              = 1.0f / sqrt(r2);
+                    float invR              = 1.0f / sqrtf(r2);
                     float sig               = a.x + psA[j].sig;
                     float sig2              = invR * sig;
                     sig2                   *= sig2;
@@ -198,9 +198,9 @@ void METHOD_NAME(kCalculateCDLJObcGbsa, Forces1_kernel)(unsigned int* workUnit )
                     // ObcGbsaForce1 part
                     float alpha2_ij         = br * psA[j].br;
                     float D_ij              = r2 / (4.0f * alpha2_ij);
-                    float expTerm           = exp(-D_ij);
+                    float expTerm           = expf(-D_ij);
                     float denominator2      = r2 + alpha2_ij * expTerm;
-                    float denominator       = sqrt(denominator2);
+                    float denominator       = sqrtf(denominator2);
                     float Gpol              = (q2 * psA[j].q) / (denominator * denominator2);
                     float dGpol_dalpha2_ij  = -0.5f * Gpol * expTerm * (1.0f + D_ij);
                     dEdR                   += Gpol * (1.0f - 0.25f * expTerm);
@@ -283,14 +283,14 @@ void METHOD_NAME(kCalculateCDLJObcGbsa, Forces1_kernel)(unsigned int* workUnit )
                         float dy                = psA[tj].y - apos.y;
                         float dz                = psA[tj].z - apos.z;
 #ifdef USE_PERIODIC
-                        dx                     -= floor(dx*cSim.invPeriodicBoxSizeX+0.5f)*cSim.periodicBoxSizeX;
-                        dy                     -= floor(dy*cSim.invPeriodicBoxSizeY+0.5f)*cSim.periodicBoxSizeY;
-                        dz                     -= floor(dz*cSim.invPeriodicBoxSizeZ+0.5f)*cSim.periodicBoxSizeZ;
+                        dx                     -= floorf(dx*cSim.invPeriodicBoxSizeX+0.5f)*cSim.periodicBoxSizeX;
+                        dy                     -= floorf(dy*cSim.invPeriodicBoxSizeY+0.5f)*cSim.periodicBoxSizeY;
+                        dz                     -= floorf(dz*cSim.invPeriodicBoxSizeZ+0.5f)*cSim.periodicBoxSizeZ;
 #endif
                         float r2                = dx * dx + dy * dy + dz * dz;
 
                         // CDLJ part
-                        float invR              = 1.0f / sqrt(r2);
+                        float invR              = 1.0f / sqrtf(r2);
                         float sig               = a.x + psA[tj].sig;
                         float sig2              = invR * sig;
                         sig2                   *= sig2;
@@ -312,9 +312,9 @@ void METHOD_NAME(kCalculateCDLJObcGbsa, Forces1_kernel)(unsigned int* workUnit )
                         // ObcGbsaForce1 part
                         float alpha2_ij         = br * psA[tj].br;
                         float D_ij              = r2 / (4.0f * alpha2_ij);
-                        float expTerm           = exp(-D_ij);
+                        float expTerm           = expf(-D_ij);
                         float denominator2      = r2 + alpha2_ij * expTerm;
-                        float denominator       = sqrt(denominator2);
+                        float denominator       = sqrtf(denominator2);
                         float Gpol              = (q2 * psA[tj].q) / (denominator * denominator2);
                         float dGpol_dalpha2_ij  = -0.5f * Gpol * expTerm * (1.0f + D_ij);
                         dEdR                   += Gpol * (1.0f - 0.25f * expTerm);
@@ -362,14 +362,14 @@ void METHOD_NAME(kCalculateCDLJObcGbsa, Forces1_kernel)(unsigned int* workUnit )
                             float dy                = psA[j].y - apos.y;
                             float dz                = psA[j].z - apos.z;
 #ifdef USE_PERIODIC
-                            dx -= floor(dx*cSim.invPeriodicBoxSizeX+0.5f)*cSim.periodicBoxSizeX;
-                            dy -= floor(dy*cSim.invPeriodicBoxSizeY+0.5f)*cSim.periodicBoxSizeY;
-                            dz -= floor(dz*cSim.invPeriodicBoxSizeZ+0.5f)*cSim.periodicBoxSizeZ;
+                            dx -= floorf(dx*cSim.invPeriodicBoxSizeX+0.5f)*cSim.periodicBoxSizeX;
+                            dy -= floorf(dy*cSim.invPeriodicBoxSizeY+0.5f)*cSim.periodicBoxSizeY;
+                            dz -= floorf(dz*cSim.invPeriodicBoxSizeZ+0.5f)*cSim.periodicBoxSizeZ;
 #endif
                             float r2                = dx * dx + dy * dy + dz * dz;
 
                             // CDLJ part
-                            float invR              = 1.0f / sqrt(r2);
+                            float invR              = 1.0f / sqrtf(r2);
                             float sig               = a.x + psA[j].sig;
                             float sig2              = invR * sig;
                             sig2                   *= sig2;
@@ -391,9 +391,9 @@ void METHOD_NAME(kCalculateCDLJObcGbsa, Forces1_kernel)(unsigned int* workUnit )
                             // ObcGbsaForce1 part
                             float alpha2_ij         = br * psA[j].br;
                             float D_ij              = r2 / (4.0f * alpha2_ij);
-                            float expTerm           = exp(-D_ij);
+                            float expTerm           = expf(-D_ij);
                             float denominator2      = r2 + alpha2_ij * expTerm;
-                            float denominator       = sqrt(denominator2);
+                            float denominator       = sqrtf(denominator2);
                             float Gpol              = (q2 * psA[j].q) / (denominator * denominator2);
                             float dGpol_dalpha2_ij  = -0.5f * Gpol * expTerm * (1.0f + D_ij);
                             dEdR                   += Gpol * (1.0f - 0.25f * expTerm);
@@ -486,14 +486,14 @@ void METHOD_NAME(kCalculateCDLJObcGbsa, Forces1_kernel)(unsigned int* workUnit )
                     float dy                = psA[tj].y - apos.y;
                     float dz                = psA[tj].z - apos.z;
 #ifdef USE_PERIODIC
-                    dx                     -= floor(dx*cSim.invPeriodicBoxSizeX+0.5f)*cSim.periodicBoxSizeX;
-                    dy                     -= floor(dy*cSim.invPeriodicBoxSizeY+0.5f)*cSim.periodicBoxSizeY;
-                    dz                     -= floor(dz*cSim.invPeriodicBoxSizeZ+0.5f)*cSim.periodicBoxSizeZ;
+                    dx                     -= floorf(dx*cSim.invPeriodicBoxSizeX+0.5f)*cSim.periodicBoxSizeX;
+                    dy                     -= floorf(dy*cSim.invPeriodicBoxSizeY+0.5f)*cSim.periodicBoxSizeY;
+                    dz                     -= floorf(dz*cSim.invPeriodicBoxSizeZ+0.5f)*cSim.periodicBoxSizeZ;
 #endif
                     float r2                = dx * dx + dy * dy + dz * dz;
 
                     // CDLJ part
-                    float invR              = 1.0f / sqrt(r2);
+                    float invR              = 1.0f / sqrtf(r2);
                     float sig               = a.x + psA[tj].sig;
                     float sig2              = invR * sig;
                     sig2                   *= sig2;
@@ -519,9 +519,9 @@ void METHOD_NAME(kCalculateCDLJObcGbsa, Forces1_kernel)(unsigned int* workUnit )
                     // ObcGbsaForce1 part
                     float alpha2_ij         = br * psA[tj].br;
                     float D_ij              = r2 / (4.0f * alpha2_ij);
-                    float expTerm           = exp(-D_ij);
+                    float expTerm           = expf(-D_ij);
                     float denominator2      = r2 + alpha2_ij * expTerm;
-                    float denominator       = sqrt(denominator2);
+                    float denominator       = sqrtf(denominator2);
                     float Gpol              = (q2 * psA[tj].q) / (denominator * denominator2);
                     float dGpol_dalpha2_ij  = -0.5f * Gpol * expTerm * (1.0f + D_ij);
                     dEdR                   += Gpol * (1.0f - 0.25f * expTerm);
